requirements.txt
add_parent_path==0.1.11
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
idna==3.7
iniconfig==2.0.0
joblib==1.4.2
kanilog==0.3.3
kiwisolver==1.4.5
logzero==1.7.0
matplotlib==3.9.1
mt4_hst==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pathlib2==2.3.7.post1
pillow==10.4.0
pluggy==1.5.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
stdlogging==0.16
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.2
